renamed ItemsForm to ItemForm for clairity
